pyproject.toml
[project]
name = "shelly"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "annotated-types==0.7.0",
    "anyio==4.9.0",
    "certifi==2025.1.31",
    "distro==1.9.0",
    "h11==0.14.0",
    "httpcore==1.0.8",
    "httpx==0.28.1",
    "idna==3.10",
    "jiter==0.9.0",
    "openai==1.74.0",
    "pydantic==2.11.3",
    "pydantic-core==2.33.1",
    "python-dotenv==1.1.0",
    "sniffio==1.3.1",
    "tqdm==4.67.1",
    "typing-extensions==4.13.2",
    "typing-inspection==0.4.0",
]
